generator: keep registers clean after function calls
Keep the registers clean after function calls by using arrays for
storing local variables, resulting in the need for only one register,
called the data register.  This also has several benefits: increasing
the number of possible local variables, allowing functions with multiple
arguments (due to the prologue now having a place to store state), and
increasing the possible number of blocks and functions (by reserving
regular registers for functions and blocks).

Add a test to make sure that the stack stays clean for functions in the
math library.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>