calculator: make q and Q work correctly with comparison operators
The comparison operators call do_parse, but when unwinding the stack, we
could accidentally not unwind it through macros called via the
comparison operators.  These macro calls are not considered macros for
the purpose of q and Q and so should be ignored for the purposes of
unwinding.  Update the code to contain a state variable indicating
whether we're unwinding, and if so, don't execute any code until we're
done.

Also, update the tests not to try to reuse a calculator object that has
exited.  This used to work (although it was not supported), but it no
longer does.  Add tests for all of these cases.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>