generator: unify loops and conditionals
In dc, like in assembly, loops and conditionals share much common code,
the main difference being the support for the false branch and whether
to jump back to the beginning of the loop.  Unify these two structures
so they share much code, making it easier to implement future forms of
loops (e.g. while) in the future.

Note that in order to exploit the similarity, the test condition is now
executed at the end of the loop, instead of having the loop execute
conditionally the first time.  This fixes a bug where we would execute
loops that should never have run at least once.  However, due to the
fact that conditionals are not considered executions of macros for the
purposes of Q, we must wrap the initial test in an immediately-executed
macro so that Q can exit from it correctly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>