generator: correct ambiguous parsing of lambda
Using a stabby lambda as the argument to ObjectSpace.define_finalizer
results in proper parsing, but using the lambda keyword results in the
block being passed to define_finalizer itself.  To avoid needing to use
parentheses, assign the lambda to a variable, and then pass the variable
to define_finalizer.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>