spec: add pending spec for SHA-512 constants
SHA-512 uses constants that are the first 64 bits of the fractional
parts of the cube roots of the first 80 primes (2 to 409).  Since we
know those values, use them as a test for the math library's e and l
functions, using the rule from algebra that a^x = e^(x*l(a)).

Mark the spec as pending as it does not yet pass due to incorrect
computations.  The algorithm can be fixed and the test no longer marked
as pending in a future commit.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>