spec: fix expected data for SHA-512 spec
The expected values should not have a leading zero, as dc will not print
them with one.  This wasn't noticed previously because the current l
function fails due to a lack of precision before it encounters one of
those values.  A properly implemented l function would have failed on
these assertions at a later time.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>