generator: improve register preallocation
If a loop used a comparison, it was possible for the registers for the
macros to not be preallocated properly, and then in the function
epilogue, we would attempt to pop an empty stack.  Compute the code for
a function first and then insert the preallocation code before the
generated function code.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>