spec/library: handle leading zeros differently
Some of the SHA-512 constants have leading zeros when written out as
16-character hex values.  However, dc, like any good calculator, doesn't
show leading zeros.  Simply omitting them in the %w block made RuboCop
unhappy, because it doesn't like multiple spaces, and lining up the
values is better for readability.  As a consequence, simply re-add the
trailing zeros and strip them off when iterating through the loop.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>