generator: avoid using extensions in non-debug mode
The generator is supposed to produce portable dc code that runs on any
system by default.  However, there were some uses of the r operator,
which is a non-portable GNU and FreeBSD extension.  Since, as the GNU dc
man page points out, this can be easily replaced with “SaSbLaLb”, do so
to improve portability.

Note that the debug mode still uses the n operator due to the lack of a
good alternative.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>